requirements/requirements_nlp.txt
boto3
einops
faiss-cpu
fasttext
flask_restful
ftfy
gdown
gradio
h5py
ijson
inflect
jieba
matplotlib>=3.3.2
nltk>=3.6.5
numpy
opencc
pangu
rapidfuzz
sacrebleu[ja]
sentence_transformers
